Move acceptable_bearing_error function to utils
