+++ pepys_import/core/validators/enhanced_validator.py
@@ -3,8 +3,8 @@ from math import degrees
 from pepys_import.utils.unit_utils import (
     bearing_between_two_points,
     distance_between_two_points_haversine,
+    acceptable_bearing_error,
 )
-
 from pepys_import.core.formats import unit_registry
 
 
@@ -140,33 +140,3 @@ class EnhancedValidator:
             }
         )
         return False
-
-
-def acceptable_bearing_error(bearing1, bearing2, delta):
-    """Determines if the two bearings are more than a set angle apart, allowing
-    for angles that span zero (North)
-
-    :param bearing1: The first bearing
-    :type bearing1: number (degrees)
-    :param bearing2: The second bearing
-    :type bearing2: number (degrees)
-    :param delta: The acceptable separation
-    :type delta: number (degrees)
-    """
-
-    try:
-        # Try treating it as a Quantity
-        bearing1_mag = bearing1.magnitude
-    except AttributeError:
-        # Otherwise just a normal float
-        bearing1_mag = float(bearing1)
-
-    try:
-        bearing2_mag = bearing2.magnitude
-    except AttributeError:
-        bearing2_mag = float(bearing2)
-
-    # note: compact test algorithm came from here:
-    #    https://gamedev.stackexchange.com/a/4472/8270
-    diff = 180 - abs(abs(bearing1_mag - bearing2_mag) - 180)
-    return diff <= delta
