+++ tests/test_enhanced_validator.py
@@ -4,13 +4,11 @@ from datetime import datetime, timedelta
 
 from pepys_import.core.formats import unit_registry
 from pepys_import.core.store.data_store import DataStore
-from pepys_import.core.validators.enhanced_validator import (
-    EnhancedValidator,
-    acceptable_bearing_error,
-)
+from pepys_import.core.validators.enhanced_validator import EnhancedValidator
 from pepys_import.core.validators import constants
 from pepys_import.file.importer import Importer
 from pepys_import.core.formats.location import Location
+from pepys_import.utils.unit_utils import acceptable_bearing_error
 
 
 class EnhancedValidatorTestCase(unittest.TestCase):
